pyproject.toml
[tool.poetry]
name = "chiru"
version = "0.7.0"
description = "Discord client library for AnyIO"
authors = ["Lura Skye <l@veriny.tf>"]
readme = "README.rst"
packages = [{include = "chiru", from = "src"}]
license = "LGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.0"
attrs = "^23.1.0"
stickney = {git = "https://github.com/Fuyukai/stickney.git"}
anyio = "^4.0.0"
cattrs = {git = "https://github.com/python-attrs/cattrs.git"}
arrow = "^1.3.0"
bitarray = "^2.8.2"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
respect-gitignore = true
src = ["src/chiru"]
line-length = 100

select = [
  "RUF", 
  "F", 
  "E", 
  "W",
  "I", 
  "UP", 
  "B",  
  "YTT", 
  "ASYNC",
  "TRIO", 
  "PYI", 
  "SIM", 
  "RET",
]
ignore = [
    "W291",  # SHUT THE FUCK UP WHEN I'M TYPING COMMENTS
    "W293",
]

[tool.ruff.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = "true"

local_partial_types = true
warn_unused_ignores = true
warn_unused_configs = true
warn_redundant_casts = true
warn_return_any = true


disallow_any_decorated = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
check_untyped_defs = true

[tool.poetry.group.dev.dependencies]
trio = "^0.23.0"
black = "^23.10.1"
ruff = "^0.1.4"
mypy = "^1.7.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = ">1.3"
sphinx-autodoc-typehints = "^1.25.2"
sphinxcontrib-jquery = "^4.1"
sphinx-inline-tabs = "^2023.4.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
